Provide CloudFront domain name to Lambda after deployment
Pass the CloudFront distribution domain name, which for now is
the base URL for the website, to the website's backend Lambda
function to enable Cognito auth integration.

Since the CloudFront distribution indirectly depends on the
backend Lambda function, and the Lambda function depends on the
distribution's domain name for Cognito auth, this creates a
circular dependency that we resolve via a custom resource that
updates the Lambda's env vars after both the Lambda and CloudFront
distribution have been deployed, with the caveat that the Lambda
will temporarily have invalid URL references in runtime code
until the change has deployed (potentially up to a minute per
deployment.

This is fine for a demo website but would want to resolve it if
productionalizing this by buying a custom domain name and passing
it to the Lambda function and CloudFront distribution to resolve
the underlying circular dependency.