Fix Cognito set-up, remove unnecessary custom resource
Update the website's Cognito set-up to provide the Cognito
user pool ID which is required for the NextJS/React web
app's Cognito integration.

Remove the custom resource that was intended to dynamically
update the website lambda function's env vars to pass in
the domain URL of the deployed CloudFront distribution,
which was not working and is no longer needed.